+++ give-divi.php
@@ -3,7 +3,7 @@
  * Plugin Name: Give - Donation Modules for Divi
  * Plugin URI:  https://go.givewp.com/divi-addon
  * Description: Use GiveWP shortcodes as Divi modules
- * Version:     1.0.0
+ * Version:     1.0.1
  * Author:      GiveWP
  * Author URI:  https://givewp.com/
  * Text Domain: give-divi
@@ -20,7 +20,7 @@ defined( 'ABSPATH' ) or exit;
 define( 'GIVE_DIVI_ADDON_NAME', 'Give - Divi' );
 
 // Versions
-define( 'GIVE_DIVI_ADDON_VERSION', '1.0.0' );
+define( 'GIVE_DIVI_ADDON_VERSION', '1.0.1' );
 define( 'GIVE_DIVI_ADDON_MIN_GIVE_VERSION', '2.9.6' );
 
 // Add-on paths