+++ CHANGELOG.md
@@ -5,6 +5,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+Divi Modules Settings (#40)
+
 ## 1.0.0 - 2020-01-28
 
 ### Added