+++ CHANGELOG.md
@@ -5,8 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 1.0.1 - 2020-03-30
+
 ### Fixed
-Divi Modules Settings (#40)
+
+-   Divi Modules Settings (#40)
 
 ## 1.0.0 - 2020-01-28